Introduces an dispatcher lib for the executors crate
A stage_dispatch_executors lib is introduced that provides a dispatcher for the Executors crate. I expect one of these to exist per executor lib that we intend to support.

As a consequence, I've introduced workspaces into the build so we can aggregate it.
