+++ Cargo.toml
@@ -1,14 +1,5 @@
-[package]
-name = "stage"
-version = "0.1.0"
-authors = ["huntc <huntchr@gmail.com>"]
-edition = "2018"
-
-[dependencies]
-crossbeam-channel = "0.5"
-log = "0.4"
-
-[dev-dependencies]
-async-std = "1.9.0"
-env_logger = "0.8.3"
-executors = "0.8.0"
+[workspace]
+members = [
+    "core",
+    "executors"
+]
